+++ 5.0/load.sh
@@ -6,8 +6,10 @@ set -e
 ingest
 
 # Add Kibana Dashboard
+: ${ES_URL:=http://elasticsearch:9200}
+
 if [ -z ${ES_USERNAME} ]; then
-  import_dashboards -es http://elasticsearch:9200 -dir /nginx_data/nginx-dashboard;
+  import_dashboards -es ${ES_URL} -dir /nginx_data/nginx-dashboard;
 else
-  import_dashboards -es http://elasticsearch:9200 -user $ES_USERNAME -pass $ES_PASSWORD -dir /nginx_data/nginx-dashboard;
+  import_dashboards -es ${ES_URL} -user $ES_USERNAME -pass $ES_PASSWORD -dir /nginx_data/nginx-dashboard;
 fi